models/edit-file-diff: Extract "open-editor" utility
