+++ lib/models/edit-file-diff.js
@@ -4,11 +4,11 @@ var fs           = require('fs');
 var Promise      = require('../ext/promise');
 var readFile     = Promise.denodeify(fs.readFile);
 var writeFile    = Promise.denodeify(fs.writeFile);
-var childProcess = require('child_process');
 var jsdiff       = require('diff');
 var quickTemp    = require('quick-temp');
 var path         = require('path');
 var SilentError  = require('silent-error');
+var openEditor   = require('../utilities/open-editor');
 
 function EditFileDiff(options) {
   this.info = options.info;
@@ -53,21 +53,10 @@ function invokeEditor(result) {
   var diff     = jsdiff.createPatch(info.outputPath, result.output.toString(), result.input);
   var diffPath = path.join(this.tmpDifferenceDir, 'currentDiff.diff'); // jshint ignore:line
 
-  return new Promise(function(resolve, reject) {
-    writeFile(diffPath, diff).then(function() {
-      var editorArgs  = process.env.EDITOR.split(' ');
-      var editor      = editorArgs.shift();
-      var editProcess = childProcess.spawn(editor, [diffPath].concat(editorArgs), {stdio: 'inherit'});
-      var results     = { outputPath: info.outputPath, diffPath: diffPath };
-
-      editProcess.on('close', function(code) {
-        if (code === 0) {
-          resolve(results);
-        } else {
-          reject();
-        }
-      });
-    });
+  return writeFile(diffPath, diff).then(function() {
+    return openEditor(diffPath);
+  }).then(function() {
+    return { outputPath: info.outputPath, diffPath: diffPath };
   });
 }
