+++ lib/models/file-info.js
@@ -9,6 +9,7 @@ var EditFileDiff = require('./edit-file-diff');
 var EOL          = require('os').EOL;
 var isBinaryFile = require('isbinaryfile');
 var template     = require('lodash/template');
+var canEdit      = require('../utilities/open-editor').canEdit;
 
 function processTemplate(content, context) {
   var options = {
@@ -53,10 +54,6 @@ FileInfo.prototype.confirmOverwrite = function(path) {
       return response.answer;
     });
 };
-
-function canEdit() {
-  return process.env.EDITOR !== undefined;
-}
 
 FileInfo.prototype.displayDiff = function() {
   var info = this,
